+++ src/internal/open-api.ts
@@ -2,6 +2,7 @@ import * as SchemaOpenApi from "schema-openapi"
 
 import * as AST from "@effect/schema/AST"
 import * as Schema from "@effect/schema/Schema"
+import { ReadonlyRecord } from "effect"
 import { identity, pipe } from "effect/Function"
 import * as Option from "effect/Option"
 import * as ReadonlyArray from "effect/ReadonlyArray"
@@ -98,6 +99,19 @@ export const make = (
     pathSpecs.push(SchemaOpenApi.globalTags(firstGlobalTag, ...restGlobalTags))
   }
 
+  if (api.options.servers) {
+    pathSpecs.push(
+      ...api.options.servers.map((server) =>
+        typeof server === "string"
+          ? SchemaOpenApi.server(server)
+          : SchemaOpenApi.server(
+            server.url,
+            (_s) => ({ ..._s, ...server })
+          )
+      )
+    )
+  }
+
   const openApi = SchemaOpenApi.openAPI(
     api.options.title,
     api.options.version,
@@ -125,7 +139,7 @@ const createPath = (path: string) =>
     .replace(/:(\w+)$/g, "{$1}")
 
 const descriptionSetter = <A extends { description?: string }>(
-  schema: Schema.Schema<any, any>
+  schema: Schema.Schema<any, any, any>
 ) =>
   pipe(
     schema.ast,
@@ -133,28 +147,59 @@ const descriptionSetter = <A extends { description?: string }>(
     Option.match({ onNone: () => identity<A>, onSome: SchemaOpenApi.description })
   )
 
-const createParameterSetters = (
-  type: "query" | "header" | "path",
-  schema: Schema.Schema<any>
-) => {
-  let ast = schema.ast
-
+const getPropertySignatures = (
+  openApiType: "query" | "header" | "path",
+  ast: AST.AST
+): ReadonlyArray<AST.PropertySignature> => {
   if (ast._tag === "Transform") {
-    ast = ast.from
+    return getPropertySignatures(openApiType, ast.from)
   }
 
-  if (ast._tag !== "TypeLiteral") {
-    throw new Error(
-      `${type} parameter must be a type literal schema, found ${ast._tag}`
+  if (ast._tag === "TypeLiteral") {
+    return ast.propertySignatures
+  }
+
+  if (ast._tag === "Union") {
+    const requiredPropertyNamesInUnions = pipe(
+      ast.types,
+      ReadonlyArray.map((type) =>
+        getPropertySignatures(openApiType, type).filter((ps) => !ps.isOptional).map((ps) => ps.name)
+      )
+    )
+
+    return pipe(
+      ast.types,
+      ReadonlyArray.flatMap((type) => getPropertySignatures(openApiType, type)),
+      ReadonlyArray.groupBy((ps) => String(ps.name)),
+      ReadonlyRecord.toEntries,
+      ReadonlyArray.map(([_, ps]) => {
+        const [first, ...rest] = ps
+        return AST.createPropertySignature(
+          first.name,
+          AST.createUnion(ps.map((ps) => ps.type)),
+          !requiredPropertyNamesInUnions.every((names) => names.includes(first.name)),
+          rest.reduce((acc, cur) => acc || cur.isReadonly, first.isReadonly),
+          first.annotations
+        )
+      })
     )
   }
 
-  return ast.propertySignatures.map((ps) => {
+  throw new Error(
+    `${ast._tag} is not supported for ${openApiType} parameter.`
+  )
+}
+
+const createParameterSetters = (
+  type: "query" | "header" | "path",
+  schema: Schema.Schema<any, any>
+) => {
+  return getPropertySignatures(type, schema.ast).map((ps) => {
     if (typeof ps.name !== "string") {
       throw new Error(`${type} parameter struct fields must be strings`)
     }
 
-    const schema = Schema.make(ps.type)
+    const schema = Schema.make<never, unknown, unknown>(ps.type)
 
     return SchemaOpenApi.parameter(
       ps.name,
@@ -166,7 +211,7 @@ const createParameterSetters = (
   })
 }
 
-const createResponseHeadersSchemaMap = (schema: Schema.Schema<any>) => {
+const createResponseHeadersSchemaMap = (schema: Schema.Schema<any, any>) => {
   let ast = schema.ast
 
   if (ast._tag === "Transform") {
@@ -180,7 +225,7 @@ const createResponseHeadersSchemaMap = (schema: Schema.Schema<any>) => {
   return Object.fromEntries(
     ast.propertySignatures.map((ps) => [
       ps.name,
-      Schema.make<string, unknown>(ps.type)
+      Schema.make<never, string, unknown>(ps.type)
     ])
   )
 }